[all] Refactor time.Time to xtime.UnixNano
This PR refactors the usages of `time.Time` to `xtime.UnixNano` along the
critical read path primarily, to reduce overhead from constant conversions
to/from time.Times.

Benchmarks show this is around a 13% improvement.

On real-data benchmarks running `read_data_files`:

With optimization:
```
Running time: 283.808698ms
35238 series read
(124161.10 series/second)
2535668 datapoints decoded
(8934426.67 datapoints/second)
```

Without optimization:
```
35238 series read
(109477.95 series/second)

2535668 datapoints decoded
(7877851.54 datapoints/second)
```

Still TODO: update test files, get them running

+++ src/cmd/services/m3coordinator/ingest/write.go
@@ -280,7 +280,8 @@ func (d *downsamplerAndWriter) writeToDownsampler(
 		if result.ShouldDropTimestamp {
 			err = result.SamplesAppender.AppendUntimedGaugeSample(dp.Value, annotation)
 		} else {
-			err = result.SamplesAppender.AppendGaugeSample(dp.Timestamp, dp.Value, annotation)
+			err = result.SamplesAppender.AppendGaugeSample(int64(dp.Timestamp),
+				dp.Value, annotation)
 		}
 		if err != nil {
 			return result.IsDropPolicyApplied, err
@@ -507,19 +508,23 @@ func (d *downsamplerAndWriter) writeAggregatedBatch(
 				if result.ShouldDropTimestamp {
 					err = result.SamplesAppender.AppendUntimedGaugeSample(dp.Value, value.Annotation)
 				} else {
-					err = result.SamplesAppender.AppendGaugeSample(dp.Timestamp, dp.Value, value.Annotation)
+					err = result.SamplesAppender.AppendGaugeSample(int64(dp.Timestamp),
+						dp.Value, value.Annotation)
 				}
 			case ts.M3MetricTypeCounter:
 				if result.ShouldDropTimestamp {
-					err = result.SamplesAppender.AppendUntimedCounterSample(int64(dp.Value), value.Annotation)
+					err = result.SamplesAppender.AppendUntimedCounterSample(
+						int64(dp.Value), value.Annotation)
 				} else {
-					err = result.SamplesAppender.AppendCounterSample(dp.Timestamp, int64(dp.Value), value.Annotation)
+					err = result.SamplesAppender.AppendCounterSample(int64(dp.Timestamp),
+						int64(dp.Value), value.Annotation)
 				}
 			case ts.M3MetricTypeTimer:
 				if result.ShouldDropTimestamp {
 					err = result.SamplesAppender.AppendUntimedTimerSample(dp.Value, value.Annotation)
 				} else {
-					err = result.SamplesAppender.AppendTimerSample(dp.Timestamp, dp.Value, value.Annotation)
+					err = result.SamplesAppender.AppendTimerSample(int64(dp.Timestamp),
+						dp.Value, value.Annotation)
 				}
 			}
 			if err != nil {
